Minor casting fix

+++ ASP2/DZ3/AddressFunction.cpp
@@ -8,7 +8,7 @@ void AddressFunction::init(StaticHashTable* hashTable) const {
 	this->hashTable = hashTable;
 }
 
-unsigned int SplitSequenceLinearHashing::getAddress(unsigned int key, int address, int attempt, int size) const {
+unsigned int SplitSequenceLinearHashing::getAddress(unsigned int key, int address, int attempt, size_t size) const {
 	auto toCompare = hashTable->table[address].front();
 
 	auto movement = toCompare->getId() > key ? s1 : s2;
