+++ ASP2/DZ3/AddressFunction.hpp
@@ -8,8 +8,8 @@ protected:
 public:
 	void init(StaticHashTable* hashTable) const;
 
-	virtual unsigned int getAddress(unsigned int key, int address, int attempt, int size) const = 0;
-	unsigned int operator()(unsigned int key, int address, int attempt, int size) const { return getAddress(key, address, attempt, size); }
+	virtual unsigned int getAddress(unsigned int key, int address, int attempt, size_t size) const = 0;
+	unsigned int operator()(unsigned int key, int address, int attempt, size_t size) const { return getAddress(key, address, attempt, size); }
 };
 
 class SplitSequenceLinearHashing : public AddressFunction {
@@ -18,5 +18,5 @@ private:
 public:
 	SplitSequenceLinearHashing(int s1, int s2) : s1(s1), s2(s2) {}
 
-	unsigned int getAddress(unsigned int key, int address, int attempt, int size) const override;
+	unsigned int getAddress(unsigned int key, int address, int attempt, size_t size) const override;
 };
